Struct alga::general::Id

#[repr(C)]
pub struct Id<O: Operator = Multiplicative> { /* private fields */ }

The universal identity element wrt. a given operator, usually noted Id with a context-dependent subscript.

By default, it is the multiplicative identity element. It represents the degenerate set containing only the identity element of any group-like structure. It has no dimension known at compile-time. All its operations are no-ops.

Implementations

impl<O: Operator> Id<O>

pub fn new() -> Id<O>

Creates a new identity element.