Trait alga::general::AbstractField

pub trait AbstractField<A: Operator = Additive, M: Operator = Multiplicative>: AbstractRingCommutative<A, M> + AbstractGroupAbelian<M> { }

A field is a commutative ring, and an Abelian group under both operators.

A field is a set with two binary operations, an addition and a multiplication, which are both closed, commutative, associative possess the divisibility property and an identity element, noted 0 and 1 respectively. Furthermore the multiplication is distributive over the addition.