Trait alga::general::AbstractGroup

pub trait AbstractGroup<O: Operator>: AbstractLoop<O> + AbstractMonoid<O> { }

A group is a loop and a monoid at the same time.

A groups is a set with a closed associative binary operation with the divisibility property and an identity element.