Trait alga::general::AbstractLoop

pub trait AbstractLoop<O: Operator>: AbstractQuasigroup<O> + Identity<O> { }

A loop is a quasigroup with an unique identity element, e.

A set equipped with a closed binary operation possessing the divisibility property and a unique identity element.

Identity element

∃! e ∈ Self, ∀ a ∈ Self, ∃ r, l ∈ Self such that l ∘ a = a ∘ r = e.

The left inverse r and right inverse l are not required to be equal.

This property follows from

∀ a ∈ Self, ∃ e ∈ Self, such that e ∘ a = a ∘ e = a.