Trait alga::general::AbstractMagma

pub trait AbstractMagma<O: Operator>: Sized + Clone {
    // Required method
    fn operate(&self, right: &Self) -> Self;

    // Provided method
    fn op(&self, _: O, lhs: &Self) -> Self { ... }
}

A magma is an algebraic structure which consists of a set equipped with a binary operation, ∘, which must be closed.

Closed binary operation

a, b ∈ Self ⇒ a ∘ b ∈ Self

Required Methods

fn operate(&self, right: &Self) -> Self

Performs an operation.

Provided Methods

fn op(&self, _: O, lhs: &Self) -> Self

Performs specific operation.

Object Safety

This trait is not object safe.