Trait alga::general::AbstractRingCommutative

pub trait AbstractRingCommutative<A: Operator = Additive, M: Operator = Multiplicative>: AbstractRing<A, M> {
    fn prop_mul_is_commutative_approx(args: (Self, Self)) -> bool where Self: ApproxEq { ... }
    fn prop_mul_is_commutative(args: (Self, Self)) -> bool where Self: Eq { ... }
}

A ring with a commutative multiplication.

∀ a, b ∈ Self, a × b = b × a

Provided Methods

fn prop_mul_is_commutative_approx(args: (Self, Self)) -> bool where Self: ApproxEq

Returns true if the multiplication operator is commutative for the given argument tuple. Approximate equality is used for verifications.

fn prop_mul_is_commutative(args: (Self, Self)) -> bool where Self: Eq

Returns true if the multiplication operator is commutative for the given argument tuple.