Trait alga::general::AbstractSemigroup

pub trait AbstractSemigroup<O: Operator>: PartialEq + AbstractMagma<O> {
    fn prop_is_associative_approx(args: (Self, Self, Self)) -> bool
    where
        Self: RelativeEq,
    { ... }

    fn prop_is_associative(args: (Self, Self, Self)) -> bool
    where
        Self: Eq,
    { ... }
}

A semigroup is a quasigroup that is associative.

A semigroup is a set equipped with a closed associative binary operation and that has the divisibility property.

Associativity

∀ a, b, c ∈ Self, (a ∘ b) ∘ c = a ∘ (b ∘ c)

Provided methods

fn prop_is_associative_approx(args: (Self, Self, Self)) -> bool where
Self: RelativeEq,

Returns true if associativity holds for the given arguments. Approximate equality is used for verifications.

fn prop_is_associative(args: (Self, Self, Self)) -> bool where
Self: Eq,

Returns true if associativity holds for the given arguments.