Trait alga::general::Inverse

pub trait Inverse<O: Operator>: Sized {
    fn inverse(&self) -> Self;

    fn inverse_mut(&mut self) { ... }
}

Trait used to define the inverse element relative to the given operator.

The operator, e.g., Multiplicative or Additive, is identified by the type parameter O.

Required Methods

fn inverse(&self) -> Self

Returns the inverse of self, relative to the operator O.

Provided Methods

fn inverse_mut(&mut self)

In-place inversin of self.