Trait alga::general::TwoSidedInverse

pub trait TwoSidedInverse<O: Operator>: Sized {
    // Required method
    fn two_sided_inverse(&self) -> Self;

    // Provided method
    fn two_sided_inverse_mut(&mut self) { ... }
}

Trait used to define the two_sided_inverse element relative to the given operator.

The operator, e.g., Additive or Multiplicative, is identified by the type parameter O.

Required Methods

fn two_sided_inverse(&self) -> Self

Returns the two_sided_inverse of self, relative to the operator O.

The parameter O is generally either Additive or Multiplicative.

Provided Methods

fn two_sided_inverse_mut(&mut self)

In-place inversion of self, relative to the operator O.

The parameter O is generally either Additive or Multiplicative.

Object Safety

This trait is not object safe.