Trait alga::linear::InnerSpace

pub trait InnerSpace: NormedSpace {
    // Required method
    fn inner_product(&self, other: &Self) -> Self::ComplexField;

    // Provided method
    fn angle(&self, other: &Self) -> Self::RealField { ... }
}

A vector space equipped with an inner product.

It must be a normed space as well and the norm must agree with the inner product. The inner product must be symmetric, linear in its first argument, and positive definite.

Required Methods

fn inner_product(&self, other: &Self) -> Self::ComplexField

Computes the inner product of self with other.

Provided Methods

fn angle(&self, other: &Self) -> Self::RealField

Measures the angle between two vectors.

Object Safety

This trait is not object safe.