Trait alga::linear::MatrixMut

pub trait MatrixMut: Matrix {
    // Required methods
    fn set_row_mut(&mut self, i: usize, row: &Self::Row);
    fn set_column_mut(&mut self, i: usize, col: &Self::Column);
    unsafe fn set_unchecked(&mut self, i: usize, j: usize, val: Self::Field);

    // Provided methods
    fn set_row(&self, i: usize, row: &Self::Row) -> Self { ... }
    fn set_column(&self, i: usize, col: &Self::Column) -> Self { ... }
    fn set(&mut self, i: usize, j: usize, val: Self::Field) { ... }
}

The space of all matrices that are stable under modifications of its components, rows and columns.

Required Methods

fn set_row_mut(&mut self, i: usize, row: &Self::Row)

In-place sets the i-th row of this matrix.

fn set_column_mut(&mut self, i: usize, col: &Self::Column)

In-place sets the i-th col of this matrix.

unsafe fn set_unchecked(&mut self, i: usize, j: usize, val: Self::Field)

Sets the component at row i and column j of this matrix without bound checking.

Provided Methods

fn set_row(&self, i: usize, row: &Self::Row) -> Self

Sets the i-th row of this matrix.

fn set_column(&self, i: usize, col: &Self::Column) -> Self

Sets the i-th col of this matrix.

fn set(&mut self, i: usize, j: usize, val: Self::Field)

Sets the component at row i and column j of this matrix.

Object Safety

This trait is not object safe.