Trait alga::linear::ProjectiveTransformation

pub trait ProjectiveTransformation<E: EuclideanSpace>: MultiplicativeGroup + Transformation<E> {
    fn inverse_transform_point(&self, pt: &E) -> E;

    fn inverse_transform_vector(&self, v: &E::Coordinates) -> E::Coordinates;
}

The most general form of inversible transformations on an euclidean space.

Required Methods

fn inverse_transform_point(&self, pt: &E) -> E

Applies this group's inverse action on a point from the euclidean space.

fn inverse_transform_vector(&self, v: &E::Coordinates) -> E::Coordinates

Applies this group's inverse action on a vector from the euclidean space.

If v is a vector and a, b two point such that v = a - b, the action ∘ on a vector is defined as self ∘ v = (self × a) - (self × b).