Trait alga::general::AbstractMagma

pub trait AbstractMagma<O: Operator>: Sized + Clone {
    fn operate(&self, right: &Self) -> Self;

    fn op(&self, _: O, lhs: &Self) -> Self { ... }
}

A magma is an algebraic structure which consists of a set equipped with a binary operation, ∘, which must be closed.

Closed binary operation

a, b ∈ Self ⇒ a ∘ b ∈ Self

Required methods

fn operate(&self, right: &Self) -> Self

Performs an operation.

Loading content...
Provided methods

fn op(&self, _: O, lhs: &Self) -> Self

Performs specific operation.

Loading content...