Trait alga::general::AbstractMonoid

pub trait AbstractMonoid<O: Operator>: AbstractSemigroup<O> + Identity<O> {
    fn prop_operating_identity_element_is_noop_approx(args: (Self,)) -> bool
    where
        Self: ApproxEq,
    { ... }

    fn prop_operating_identity_element_is_noop(args: (Self,)) -> bool
    where
        Self: Eq,
    { ... }
}

A semigroup equipped with an identity element.

∃ e ∈ Self, ∀ a ∈ Self, e ∘ a = a ∘ e = a

Provided Methods

fn prop_operating_identity_element_is_noop_approx(args: (Self,)) -> bool where
Self: ApproxEq,

Checks whether operating with the identity element is a no-op for the given argument. Approximate equality is used for verifications.

fn prop_operating_identity_element_is_noop(args: (Self,)) -> bool where
Self: Eq,

Checks whether operating with the identity element is a no-op for the given argument.