Trait alga::general::TwoSidedInverse

pub trait TwoSidedInverse<O: Operator>: Sized {
    fn two_sided_inverse(&self) -> Self;

    fn two_sided_inverse_mut(&mut self) { ... }
}

Trait used to define the two_sided_inverse element relative to the given operator.

The operator, e.g., Additive or Multiplicative, is identified by the type parameter O.

Required methods

fn two_sided_inverse(&self) -> Self

Returns the two_sided_inverse of self, relative to the operator O.

The parameter O is generally either Additive or Multiplicative.

Loading content...
Provided methods

fn two_sided_inverse_mut(&mut self)

In-place inversion of self, relative to the operator O.

The parameter O is generally either Additive or Multiplicative.

Loading content...