Module alga::linear

Traits dedicated to linear algebra.

Traits

AffineSpace
A set points associated with a vector space and a transitive and free additive group action (the translation).

AffineTransformation
The group of affine transformations. They are decomposable into a rotation, a non-uniform scaling, a second rotation, and a translation (applied in that order).

DirectIsometry
Subgroups of the orientation-preserving isometry group SE(n), i.e., rotations and translations.

EuclideanSpace
The finite-dimensional affine space based on the field of reals.

FiniteDimInnerSpace
A finite-dimensional vector space equipped with an inner product that must coincide with the dot product.

FiniteDimVectorSpace
A finite-dimensional vector space.

InnerSpace
A vector space equipped with an inner product.

InversibleSquareMatrix
The group of inversible matrix. Commonly known as the General Linear group GL(n) by algebraists.

Isometry
Subgroups of the isometry group E(n), i.e., rotations, reflexions, and translations.

Matrix
The space of all matrices.

MatrixMut
The space of all matrices that are stable under modifications of its components, rows and columns.

NormedSpace
A normed vector space.

OrthogonalTransformation
Subgroups of the n-dimensional rotations and scaling O(n).

ProjectiveTransformation
The most general form of invertible transformations on an euclidean space.

Rotation
Subgroups of the n-dimensional rotation group SO(n).

Scaling
Subgroups of the (signed) uniform scaling group.

Similarity
Subgroups of the similarity group S(n), i.e., rotations, translations, and (signed) uniform scaling.

SquareMatrix
The monoid of all square matrices, including non-inversible ones.

SquareMatrixMut
The monoid of all mutable square matrices that are stable under modification of its diagonal.

Transformation
A general transformation acting on an euclidean space. It may not be inversible.

Translation
Subgroups of the n-dimensional translation group T(n).

VectorSpace
A vector space has a module structure over a field instead of a ring.