Trait alga::linear::AffineSpace

pub trait AffineSpace: Sized + Clone + PartialEq + Sub<Self, Output = Self::Translation> + ClosedAdd<Self::Translation> {
    type Translation: VectorSpace;
    fn translate_by(&self, t: &Self::Translation) -> Self { ... }

    fn subtract(&self, right: &Self) -> Self::Translation { ... }
}

A set points associated with a vector space and a transitive and free additive group action (the translation).

Associated Types

type Translation: VectorSpace

The associated vector space.

Provided methods

fn translate_by(&self, t: &Self::Translation) -> Self

Same as *self + *t. Applies the additive group action of this affine space’s associated vector space on self.

fn subtract(&self, right: &Self) -> Self::Translation

Same as *self - *other. Returns the unique element v of the associated vector space such that self = right + v.