Trait alga::linear::InnerSpace

pub trait InnerSpace: NormedSpace<Field=Self::Real> {
    type Real: Real;
    fn inner_product(&self, other: &Self) -> Self::Real;

    fn angle(&self, other: &Self) -> Self::Real { ... }
}

A vector space aquipped with an inner product.

It must be a normed space as well and the norm must agree with the inner product. The inner product must be symmetric, linear in its first agurment, and positive definite.

Associated Types

type Real: Real

The result of inner product (same as the field used by this vector space).

Required Methods

fn inner_product(&self, other: &Self) -> Self::Real

Computes the inner product of self with other.

Provided Methods

fn angle(&self, other: &Self) -> Self::Real

Measures the angle between two vectors.