Trait alga::linear::NormedSpace

pub trait NormedSpace: VectorSpace {
    fn norm_squared(&self) -> Self::Field;
    fn norm(&self) -> Self::Field;
    fn normalize(&self) -> Self;
    fn normalize_mut(&mut self) -> Self::Field;
    fn try_normalize(&self, eps: Self::Field) -> Option<Self>;
    fn try_normalize_mut(&mut self, eps: Self::Field) -> Option<Self::Field>;
}

A normed vector space.

Required Methods

fn norm_squared(&self) -> Self::Field

The squared norm of this vector.

fn norm(&self) -> Self::Field

The norm of this vector.

fn normalize(&self) -> Self

Returns a normalized version of this vector.

fn normalize_mut(&mut self) -> Self::Field

Normalizes this vector in-place and returns its norm.

fn try_normalize(&self, eps: Self::Field) -> Option<Self>

Returns a normalized version of this vector unless its norm as smaller or equal to eps.

fn try_normalize_mut(&mut self, eps: Self::Field) -> Option<Self::Field>

Normalizes this vector in-place or does nothing if its norm is smaller or equal to eps.

If the normalization succeded, returns the old normal of this vector.