Trait alga::linear::Scaling

pub trait Scaling<E: EuclideanSpace>: AffineTransformation<E, NonUniformScaling=Self, Translation=Id, Rotation=Id> + SubsetOf<E::Real> {
    fn to_real(&self) -> E::Real { ... }
    fn from_real(r: E::Real) -> Option<Self> { ... }
    fn powf(&self, n: E::Real) -> Option<Self> { ... }
    fn scale_between(a: &E::Coordinates, b: &E::Coordinates) -> Option<Self> { ... }
}

Subgroups of the (signed) uniform scaling group.

Provided Methods

fn to_real(&self) -> E::Real

Converts this scaling factor to a real. Same as self.to_superset().

fn from_real(r: E::Real) -> Option<Self>

Attempts to convert a real to an element of this scaling subgroup. Same as Self::from_superset(). Returns None if no such scaling is possible for this subgroup.

fn powf(&self, n: E::Real) -> Option<Self>

Raises the scaling to a power. The result must be equivalent to self.to_superset().powf(n). Returns None if the result is not representable by Self.

fn scale_between(a: &E::Coordinates, b: &E::Coordinates) -> Option<Self>

The scaling required to make a have the same norm as b, i.e., |b| = |a| * norm_ratio(a, b).